feat: enforce trusted-contact ACL for inbound voice + extract shared access-request helper
