+++ assistant/src/calls/relay-server.ts
@@ -13,6 +13,8 @@ import type { ServerWebSocket } from 'bun';
 import { getConfig } from '../config/loader.js';
 import * as conversationStore from '../memory/conversation-store.js';
 import { revokeScopedApprovalGrantsForContext } from '../memory/scoped-approval-grants.js';
+import { notifyGuardianOfAccessRequest } from '../runtime/access-request-helper.js';
+import { resolveActorTrust, toGuardianContextCompat } from '../runtime/actor-trust-resolver.js';
 import {
   getPendingChallenge,
   validateAndConsumeChallenge,
@@ -471,6 +473,71 @@ export class RelayConnection {
     if (!isInbound && verificationConfig.enabled) {
       await this.startVerification(session, verificationConfig);
     } else if (isInbound) {
+      // ── Trusted-contact ACL enforcement for inbound voice ──
+      // Resolve the caller's trust classification before allowing the call
+      // to proceed. Guardian and trusted-contact callers pass through;
+      // unknown callers are denied with deterministic voice copy and an
+      // access request is created for the guardian.
+      const actorTrust = resolveActorTrust({
+        assistantId,
+        sourceChannel: 'voice',
+        externalChatId: msg.from,
+        senderExternalUserId: msg.from || undefined,
+      });
+
+      if (actorTrust.trustClass === 'unknown') {
+        log.info(
+          { callSessionId: this.callSessionId, from: msg.from, trustClass: actorTrust.trustClass },
+          'Inbound voice ACL: unknown caller denied',
+        );
+
+        recordCallEvent(this.callSessionId, 'inbound_acl_denied', {
+          from: msg.from,
+          trustClass: actorTrust.trustClass,
+          denialReason: actorTrust.denialReason,
+        });
+
+        // Create/dedupe canonical access request via the shared helper
+        try {
+          notifyGuardianOfAccessRequest({
+            canonicalAssistantId: assistantId,
+            sourceChannel: 'voice',
+            externalChatId: msg.from,
+            senderExternalUserId: actorTrust.canonicalSenderId ?? msg.from,
+          });
+        } catch (err) {
+          log.error({ err, callSessionId: this.callSessionId }, 'Failed to create access request for denied voice caller');
+        }
+
+        // Deny with deterministic voice copy and end the call
+        this.sendTextToken(
+          'This number is not authorized. Your request has been forwarded to the account guardian.',
+          true,
+        );
+
+        updateCallSession(this.callSessionId, {
+          status: 'failed',
+          endedAt: Date.now(),
+          lastError: 'Inbound voice ACL: caller not authorized',
+        });
+
+        setTimeout(() => {
+          this.endSession('Inbound voice ACL denied');
+        }, 3000);
+        return;
+      }
+
+      // Guardian and trusted-contact callers proceed normally.
+      // Update the controller's guardian context with the trust-resolved
+      // context so downstream policy gates have accurate actor metadata.
+      if (this.controller) {
+        const resolvedGuardianContext = toGuardianRuntimeContext(
+          'voice',
+          toGuardianContextCompat(actorTrust, msg.from),
+        );
+        this.controller.setGuardianContext(resolvedGuardianContext);
+      }
+
       // For inbound calls, check if there's a pending voice guardian
       // challenge that the caller needs to complete before proceeding.
       const pendingChallenge = getPendingChallenge(assistantId, 'voice');
